+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [[*next-version*]] - YYYY-MM-DD
 
 ## [0.1.1-alpha3] - 2023-02-01
-### Changed
+### Added
 - The `Constructor`, `FuncService`, `ServiceList`, `StringService`, `Extension`, and `ArrayExtension` services now
   accept other service instances as dependencies, in addition to strings (#14).
+
+### Deprecated
 - The `ResolveKeysCabableTrait::resolveKeys()` method is now deprecated. Clients should use the `resolveDeps()` method
   instead (#14).